app: extract newAppCRToUpdate to a separate file
This function will be the only real difference between reusable
resources. Other stuff is just a copy paste + replacements. Something we
can potentially generate. newOjbectTypeToUpdate will be custom for each
object and needs different testing. I'd like to extract it to make it
easier to create and review new resources.

Creating a resource would boil down to those three tasks:

1. Copy one of existing generic resources.
2. Update imports and log messages. E.g. s/App CR/ConfigMap/,
   s/appCR/configMap/, etc.
3. Replace newAppCRToUpdate with newConfigMapToUpdate (this is the only
   place where semantic has to be changed).
